+++ ghost/admin/app/components/members/modals/bulk-add-label.js
