+++ ghost/admin/app/components/members/modals/bulk-delete.js
