+++ ghost/admin/app/components/members/modals/bulk-remove-label.hbs
