+++ ghost/admin/app/controllers/members.js
@@ -1,7 +1,7 @@
-import BulkAddMembersLabelModal from '../components/modals/members/bulk-add-label';
-import BulkDeleteMembersModal from '../components/modals/members/bulk-delete';
-import BulkRemoveMembersLabelModal from '../components/modals/members/bulk-remove-label';
-import BulkUnsubscribeMembersModal from '../components/modals/members/bulk-unsubscribe';
+import BulkAddMembersLabelModal from '../components/members/modals/bulk-add-label';
+import BulkDeleteMembersModal from '../components/members/modals/bulk-delete';
+import BulkRemoveMembersLabelModal from '../components/members/modals/bulk-remove-label';
+import BulkUnsubscribeMembersModal from '../components/members/modals/bulk-unsubscribe';
 import Controller from '@ember/controller';
 import ghostPaths from 'ghost-admin/utils/ghost-paths';
 import moment from 'moment';
